moved ProcessEdgeTransitions, GetStopImpact, IsCyclewayUturn, IsPencilPointUturn, and enhancer_stats to util in mjolnir
